refactor: refactor formik checkbox fields group
refactor: refactor formik checkbox fields group

fix: fix UI issues
